controller/shows/controller_test.go
package shows

import (
	"MovieBooking/mocks"
	"MovieBooking/models"
	"MovieBooking/repository"
	"bytes"
	"encoding/json"
	"net/http"
	"net/http/httptest"
	"testing"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/mock"
)

var rPath = "/test"

func TestShowController_AddShow(t *testing.T) {
	tests := []struct {
		name string
		body models.Show
		mock func() repository.RepositoryI
		want int
	}{{
		name: "pass",
		body: models.Show{
			DateTime:  time.Now().String(),
			TheaterID: "1234",
			MovieID:   "xyz",
		},
		mock: func() repository.RepositoryI {
			ri := mocks.RepositoryI{}
			ri.On("Create", mock.Anything).Return(nil)
			return &ri
		},
		want: http.StatusCreated,
	},
		{
			name: "failes for empty Show name",
			want: http.StatusBadRequest,
			mock: func() repository.RepositoryI {
				ri := mocks.RepositoryI{}
				ri.On("Create", mock.Anything).Return(nil)
				return &ri
			},
			body: models.Show{},
		},
	}
	router := gin.Default()
	m := &ShowController{}
	router.GET(rPath, m.AddShow)
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			m.Repository = tt.mock()
			b, _ := json.Marshal(tt.body)
			req, _ := http.NewRequest("GET", rPath, bytes.NewReader(b))
			w := httptest.NewRecorder()
			router.ServeHTTP(w, req)
			t.Logf("status: %d", w.Code)
			t.Logf("response: %s", w.Body.String())
			assert.Equal(t, tt.want, w.Code)
		})
	}
}

func TestShowcontroller_GetShows(t *testing.T) {
	tests := []struct {
		name string
		body models.Show
		mock func() repository.RepositoryI
		want int
	}{{
		name: "pass",
		body: models.Show{ID: "test"},
		mock: func() repository.RepositoryI {
			ri := mocks.RepositoryI{}
			ri.On("QueryAll", mock.Anything).Return(nil)
			return &ri
		},
		want: http.StatusOK,
	}}

	router := gin.Default()
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			m := &ShowController{Repository: tt.mock()}
			router.GET(rPath, m.GetShows)
			b, _ := json.Marshal(tt.body)
			req, _ := http.NewRequest("GET", rPath, bytes.NewReader(b))
			w := httptest.NewRecorder()
			router.ServeHTTP(w, req)
			t.Logf("status: %d", w.Code)
			t.Logf("response: %s", w.Body.String())
			assert.Equal(t, tt.want, w.Code)
		})
	}
}

func TestShowcontroller_UpdateShow(t *testing.T) {
	tests := []struct {
		name string
		body models.Show
		mock func() repository.RepositoryI
		want int
	}{{
		name: "pass",
		body: models.Show{
			ID: "id",
			//Name: "Titanic",
		},
		mock: func() repository.RepositoryI {
			ri := mocks.RepositoryI{}
			ri.On("Update", mock.Anything).Return(nil)
			return &ri
		},
		want: http.StatusOK,
	}}

	router := gin.Default()
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			m := &ShowController{Repository: tt.mock()}
			router.GET(rPath, m.UpdateShow)
			b, _ := json.Marshal(tt.body)
			req, _ := http.NewRequest("GET", rPath, bytes.NewReader(b))
			w := httptest.NewRecorder()
			router.ServeHTTP(w, req)
			t.Logf("status: %d", w.Code)
			t.Logf("response: %s", w.Body.String())
			assert.Equal(t, tt.want, w.Code)
		})
	}
}

func TestShowcontroller_DeleteShow(t *testing.T) {

	tests := []struct {
		name string
		body models.Show
		mock func() repository.RepositoryI
		want int
	}{{
		name: "pass",
		body: models.Show{
			ID: "id",
			//Name: "Titanic",
		},
		mock: func() repository.RepositoryI {
			ri := mocks.RepositoryI{}
			ri.On("Delete", mock.Anything).Return(nil)
			return &ri
		},
		want: http.StatusOK,
	}}

	router := gin.Default()
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			m := &ShowController{Repository: tt.mock()}
			router.GET(rPath, m.DeleteShow)
			b, _ := json.Marshal(tt.body)
			req, _ := http.NewRequest("GET", rPath, bytes.NewReader(b))
			w := httptest.NewRecorder()
			router.ServeHTTP(w, req)
			t.Logf("status: %d", w.Code)
			t.Logf("response: %s", w.Body.String())
			assert.Equal(t, tt.want, w.Code)
		})
	}
}
